+++ server/cmd/trade/src/main/java/org/lanlance/freecartrade/model/req/RechargeRequest.java
@@ -4,11 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class RechargeRequest {
-    private BigDecimal amount;
+    private Integer amount;
 }
