+++ server/cmd/trade/src/main/java/org/lanlance/freecartrade/service/RechargeServiceIface.java
@@ -1,8 +1,5 @@
 package org.lanlance.freecartrade.service;
 
-
-import java.math.BigDecimal;
-
 public interface RechargeServiceIface {
-    void recharge(String userId, BigDecimal amount);
+    void recharge(String userId, Integer amount);
 }
